+++ postgres-types/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Added
 
+* Upgraded support for `chrono-tz` to 0.10. Use the `with-chrono-tz-0_10` feature to enable it.
+
+### Added
+
 * Added support for `chrono-tz` 0.9 via the `with-chrono-tz-0_9` feature.
 
 ## v0.2.6 - 2023-08-19